Refactor to exclude empty from params
